+++ src/module/data/actor/hero.mjs
@@ -13,7 +13,6 @@ import { systemID } from "../../constants.mjs";
  * @import ActiveEffectData from "@common/documents/_types.mjs";
  * @import AdvancementChain from "../../utils/advancement-chain.mjs";
  * @import { ActorData, ItemData } from "@common/documents/_types.mjs";
- * @import { PowerRollModifiers } from "../../_types.js";
  */
 
 const fields = foundry.data.fields;
@@ -63,8 +62,11 @@ export default class HeroModel extends CreatureModel {
       victories: requiredInteger(),
       renown: requiredInteger(),
       wealth: requiredInteger({ initial: 1 }),
-      skills: new fields.SetField(setOptions()),
       preferredKit: new fields.DocumentIdField({ readonly: false }),
+    });
+
+    schema.skills = new fields.SchemaField({
+      value: new fields.SetField(setOptions()),
     });
 
     return schema;
@@ -94,8 +96,55 @@ export default class HeroModel extends CreatureModel {
   /* -------------------------------------------------- */
 
   /** @inheritdoc */
+  static migrateData(data) {
+    foundry.abstract.Document._addDataFieldMigration(data, "hero.skills", "skills.value");
+
+    return super.migrateData(data);
+  }
+
+  /* -------------------------------------------------- */
+
+  /**
+   * Shims both source and initialized hero data.
+   * @param {object | HeroModel} data
+   */
+  static shimSkills(data) {
+    const shims = {
+      "hero.skills": "skills.value",
+      "hero.skillModifiers": "skills.modifiers",
+    };
+    for (const [oldPath, newPath] of Object.entries(shims)) {
+      // v13 version of _addDataFieldShim fails to properly safeguard dot-split properties. Results in warning spam.
+      const lastKey = oldPath.split(".").at(-1);
+      if (Object.hasOwn(data.hero, lastKey)) continue;
+      // Not using the multi-helper to improve message
+
+      let warning = `You are accessing ${this.name}#${oldPath}, which has been migrated to ${this.name}#${newPath}.`;
+
+      if (data instanceof HeroModel) {
+        warning += ` Issue occurred in Actor ${data.parent.uuid}`;
+      }
+
+      foundry.abstract.Document._addDataFieldShim(data, oldPath, newPath, { warning, since: "0.11", until: "1.0" });
+    }
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
+  static shimData(data, options) {
+    this.shimSkills(data);
+    return super.shimData(data, options);
+  }
+
+  /* -------------------------------------------------- */
+
+  /** @inheritdoc */
   prepareBaseData() {
     super.prepareBaseData();
+
+    // Existing DrawSteelActiveEffect#_applyAdd override means this also shims active effects targeting hero.skills
+    HeroModel.shimSkills(this);
 
     this.combat.initiativeThreshold = 6;
 
@@ -166,8 +215,7 @@ export default class HeroModel extends CreatureModel {
       }
     }
 
-    /** @type {Record<string, PowerRollModifiers>} */
-    this.hero.skillModifiers = { };
+    this.skills.modifiers = {};
   }
 
   /* -------------------------------------------------- */
@@ -194,11 +242,21 @@ export default class HeroModel extends CreatureModel {
 
     // Handling for trait advancements
     for (const skill of this._traits.skill ?? []) {
-      if (skill in ds.CONFIG.skills.list) this.hero.skills.add(skill);
+      if (skill in ds.CONFIG.skills.list) this.skills.value.add(skill);
     }
     for (const lang of this._traits.language ?? []) {
       if (lang in ds.CONFIG.languages) this.biography.languages.add(lang);
     }
+
+    const list = this.skills.value.reduce((skills, skill) => {
+      skill = ds.CONFIG.skills.list[skill]?.label;
+      if (skill) skills.push(skill);
+      return skills;
+    }, []).sort((a, b) => a.localeCompare(b, game.i18n.lang));
+
+    const formatter = game.i18n.getListFormatter();
+
+    this.skills.list = formatter.format(list);
   }
 
   /* -------------------------------------------------- */